Squashed commit of the following:
commit e362abfeaaed40d4e3d7d779506136efcb6d8177
Author: xupingmao <578749341@qq.com>
Date:   Tue Feb 4 23:21:42 2020 +0800

    Fix 任务中选择话题没填写到输入框的问题

commit 7e8f357add3ced94419048a60f3f5968a09f27f7
Author: xupingmao <578749341@qq.com>
Date:   Sun Feb 2 23:40:21 2020 +0800

    Fix 下拉列表不显示的问题

commit fcd8ba8ef3ded86730b9d617a35363ab22b5d084
Author: xupingmao <578749341@qq.com>
Date:   Sun Feb 2 16:11:29 2020 +0800

    优化词典

commit b0d46610cffa143ab7e7cd5c9f86a54e01e881ec
Author: xupingmao <578749341@qq.com>
Date:   Sat Feb 1 15:24:11 2020 +0800

    Update 用户界面

commit b6f642672a63dc8bdde372e6b13eff8f51b72188
Author: xupingmao <578749341@qq.com>
Date:   Fri Jan 31 18:18:18 2020 +0800

    处理置顶笔记

commit 37c91afb6c580f96f7d04486a31b735650ce3cbe
Author: xupingmao <578749341@qq.com>
Date:   Fri Jan 31 17:55:47 2020 +0800

    Update

commit e9fe64746991bb5ca4ae77e462dbdb747069f6e1
Author: xupingmao <578749341@qq.com>
Date:   Thu Jan 30 16:46:59 2020 +0800

    显示隐藏文件

commit e22737102675e2c321f992808da8db005b48b8ca
Author: xupingmao <578749341@qq.com>
Date:   Thu Jan 30 00:17:11 2020 +0800

    插件模板优化

commit 0fd12d66e99f29bfe58d4e23eb85f929b812460e
Author: xupingmao <578749341@qq.com>
Date:   Tue Jan 28 12:56:10 2020 +0800

    相册缩略图支持

commit ed76970fe841c35001468bea261156366bde6ea3
Author: xupingmao <578749341@qq.com>
Date:   Sun Jan 26 16:56:27 2020 +0800

    Fix note.get_by_name

commit 2e6642ded25407a4e9343c8b983d235e878d1586
Author: xupingmao <578749341@qq.com>
Date:   Sun Jan 26 13:05:47 2020 +0800

    代码结构重构

commit 18ddb6df9697ff6127a7e0209a727bc0ce3e2266
Author: xupingmao <578749341@qq.com>
Date:   Sat Jan 25 23:06:05 2020 +0800

    代码重构

commit 600f60d2a54b8178d2bf83ada87558e3698439d7
Author: xupingmao <578749341@qq.com>
Date:   Sat Jan 25 11:17:39 2020 +0800

    Fix 项目子项目

commit 628c68af5ed0b1c57e3fe579850358f3b57a7173
Author: xupingmao <578749341@qq.com>
Date:   Fri Jan 24 20:33:56 2020 +0800

    链接分享功能

commit 2da8e297b25602836eaf2a63b45096d21ae36d17
Author: xupingmao <578749341@qq.com>
Date:   Fri Jan 24 16:45:19 2020 +0800

    移动笔记支持搜索

commit 4e22b8bdb854031e2834c98fb6615f208f49fd5e
Author: xupingmao <578749341@qq.com>
Date:   Fri Jan 24 14:34:21 2020 +0800

    优化结构

commit 91c642f10eda92452a93ad68d624e90403e597c9
Author: xupingmao <578749341@qq.com>
Date:   Wed Jan 22 18:47:52 2020 +0800

    Fix screenshot

commit 5960b324f9fbb06d9cb5d493a099d15beff3f7a9
Author: xupingmao <578749341@qq.com>
Date:   Wed Jan 22 18:32:12 2020 +0800

    Fix login style

commit 6fcb7f4d9f7856c158b003c88b1867a28f6d0c12
Author: xupingmao <578749341@qq.com>
Date:   Wed Jan 22 12:51:26 2020 +0800

    减少内存占用

commit 562efb3a7b28eee733e50cd80a769ffe5ff3de69
Author: xupingmao <578749341@qq.com>
Date:   Wed Jan 22 12:48:56 2020 +0800

    删除无用代码，Fix笔记搜索大小写敏感的问题

commit 46fd7e1cc0bb65ed096a4e635fe2cfb03cb51ee8
Author: xupingmao <578749341@qq.com>
Date:   Tue Jan 21 23:44:15 2020 +0800

    代码优化

commit b5aa6765bc755dd09f81c2d6b838a57da15d2563
Author: xupingmao <578749341@qq.com>
Date:   Tue Jan 21 20:36:54 2020 +0800

    优化前端组件

commit 55ec7223043cebb15a86ef03f9588c41b302c974
Author: xupingmao <578749341@qq.com>
Date:   Sat Jan 18 21:38:49 2020 +0800

    优化首页性能

commit 922ae9b9ab035b9f39437bb5ab5e4471b45329e5
Author: xupingmao <578749341@qq.com>
Date:   Sat Jan 18 20:59:31 2020 +0800

    Fix 上传文件重复

commit 1e0389cf50bf0264c6ec9de81a30d91241f75311
Author: xupingmao <578749341@qq.com>
Date:   Sat Jan 18 20:36:39 2020 +0800

    优化移动端编辑

commit aba923306b759f58f08163560d71d1708da1eb60
Author: xupingmao <578749341@qq.com>
Date:   Thu Jan 16 00:15:23 2020 +0800

    Fix 项目索引更新

commit 9cb21d54dc73efb3fea1d1d35fd63859e02e2ab3
Author: xupingmao <578749341@qq.com>
Date:   Wed Jan 15 01:28:25 2020 +0800

    优化笔记搜索

commit 7f481c2c26969977d50a87363a904474e64827ef
Author: xupingmao <578749341@qq.com>
Date:   Mon Jan 13 23:43:35 2020 +0800

    更新字体大小

commit 1c5438951d8da6813b3895b02b062129210ad6c7
Author: xupingmao <578749341@qq.com>
Date:   Mon Jan 13 23:12:56 2020 +0800

    Fix iPhone无法输入中文的问题

# Conflicts:
#	README.md
#	core/xconfig.py
#	handlers/common/base_nav.html
#	handlers/note/component/create_option.html
#	handlers/note/editor/markdown_edit.html
#	handlers/note/template/note_dropdown.html
#	handlers/note/view.html
#	handlers/plugins/plugins.py

+++ handlers/note/dao.py
@@ -1,9 +1,23 @@
 # encoding=utf-8
 # Created by xupingmao on 2017/04/16
-# @modified 2020/01/12 20:54:52
+# @modified 2020/02/02 14:46:54
 
 """资料的DAO操作集合
 DAO层只做最基础的数据库交互，不做权限校验（空校验要做），业务状态检查之类的工作
+
+一些表的说明
+note_full:<note_id>              = 笔记的内容，包含一些属性（部分属性比如访问时间、访问次数不是实时更新的）
+note_index:<note_id>             = 笔记索引
+note_tiny:<user>:<note_id>       = 用户维度的笔记索引
+note_book:<user>:<note_id>       = 用户维度的笔记本(项目)索引
+token:<uuid>                     = 用于链接分享的令牌
+note_history:<note_id>:<version> = 笔记的历史版本
+note_comment:<note_id>:<timeseq> = 笔记的评论
+comment_index:<user>:<timeseq>   = 用户维度的评论索引
+search_history:<user>:<timeseq>  = 用户维度的搜索历史
+
+TODO:
+note_public:<note_id>            = 公开的笔记索引
 """
 import time
 import math
@@ -26,6 +40,8 @@ DB_PATH         = xconfig.DB_PATH
 NOTE_ICON_DICT = {
     "group"   : "fa-folder orange",
     "csv"     : "fa-table",
+    "table"   : "fa-table",
+    "post"    : "fa-file-word-o",
     "html"    : "fa-file-word-o",
     "gallery" : "fa-photo",
     "list"    : "fa-list"
@@ -195,6 +211,12 @@ def get_by_id_creator(id, creator, db=None):
         return note
     return None
 
+def get_by_token(token):
+    token_info = dbutil.get("token:%s" % token)
+    if token_info != None and token_info.type == "note":
+        return get_by_id(token_info.id)
+    return None
+
 def create_note(note_dict):
     content   = note_dict["content"]
     data      = note_dict["data"]
@@ -224,6 +246,12 @@ def create_note(note_dict):
     touch_note(parent_id)
 
     return note_id
+
+def create_token(type, id):
+    uuid = textutil.generate_uuid()
+    token_info = Storage(type = type, id = id)
+    dbutil.put("token:%s" % uuid, token_info)
+    return uuid
 
 def update_note_rank(note):
     mtime = note.mtime
@@ -258,13 +286,6 @@ def kv_put_note(note_id, note):
     # 更新索引
     update_index(note)
 
-    del_dict_key(note, "content")
-    del_dict_key(note, "data")
-    
-    if note.type == "group":
-        # 笔记本的索引
-        dbutil.put("notebook:%s:%020d" % (note.creator, int(note_id)), note)
-
 def touch_note(note_id):
     note = get_by_id(note_id)
     if note != None:
@@ -294,8 +315,11 @@ def update_index(note):
     # 更新用户索引
     dbutil.put("note_tiny:%s:%020d" % (note.creator, int(id)), note)
 
+    if note.type == "group":
+        dbutil.put("notebook:%s:%020d" % (note.creator, int(id)), note)
+
 def update_note(note_id, **kw):
-    # 移动笔记使用 move_note
+    # 这里只更新基本字段，移动笔记使用 move_note
     content   = kw.get('content')
     data      = kw.get('data')
     priority  = kw.get('priority')
@@ -306,6 +330,7 @@ def update_note(note_id, **kw):
     orderby   = kw.get("orderby")
     archived  = kw.get("archived")
     size      = kw.get("size")
+    token     = kw.get("token")
 
     old_parent_id = None
     new_parent_id = None
@@ -332,6 +357,8 @@ def update_note(note_id, **kw):
             note.archived = archived
         if size != None:
             note.size = size
+        if token != None:
+            note.token = token
 
         old_version  = note.version
         note.mtime   = xutils.format_time()
@@ -373,12 +400,12 @@ def update0(note):
     note.atime   = current_time
     kv_put_note(note.id, note)
 
-def get_by_name(name, db = None):
+def get_by_name(creator, name):
     def find_func(key, value):
         if value.is_deleted:
             return False
         return value.name == name
-    result = dbutil.prefix_list("note_tiny:", find_func, 0, 1)
+    result = dbutil.prefix_list("note_tiny:%s" % creator, find_func, 0, 1)
     if len(result) > 0:
         note = result[0]
         return get_by_id(note.id)
@@ -398,17 +425,20 @@ def delete_note(id):
     if note is None:
         return
 
-    # 复制到回收站
-    deleted_key = "note_deleted:%s:%s" % (note.creator, note.id)
-    dbutil.put(deleted_key, note)
+    if note.is_deleted != 0:
+        # 已经被删除了，执行物理删除
+        tiny_key = "note_tiny:%s:%s" % (note.creator, note.id)
+        full_key = "note_full:%s" % note.id
+        index_key = "note_index:%s" % note.id
+        dbutil.delete(tiny_key)
+        dbutil.delete(full_key)
+        dbutil.delete(index_key)
+        return
 
-    # 删除笔记
-    tiny_key = "note_tiny:%s:%s" % (note.creator, note.id)
-    full_key = "note_full:%s" % note.id
-    index_key = "note_index:%s" % note.id
-    dbutil.delete(tiny_key)
-    dbutil.delete(full_key)
-    dbutil.delete(index_key)
+    # 标记删除
+    note.mtime = xutils.format_datetime()
+    note.is_deleted = 1
+    kv_put_note(id, note)
 
     # 更新数量
     update_children_count(note.parent_id)
@@ -472,9 +502,9 @@ def list_group(creator = None, orderby = "mtime_desc", skip_archived = False):
             return False
         if skip_archived and value.parent_id != 0:
             return False
-        return value.type == "group" and value.creator == creator and value.is_deleted == 0
+        return value.type == "group" and value.is_deleted == 0
 
-    notes = dbutil.prefix_list("notebook:", list_group_func)
+    notes = dbutil.prefix_list("notebook:%s" % creator, list_group_func)
     sort_notes(notes, orderby)
     return notes
 
@@ -483,7 +513,7 @@ def list_root_group(creator = None, orderby = "name"):
     def list_root_group_func(key, value):
         return value.creator == creator and value.type == "group" and value.parent_id == 0 and value.is_deleted == 0
 
-    notes = dbutil.prefix_list("notebook:", list_root_group_func)
+    notes = dbutil.prefix_list("notebook:%s" % creator, list_root_group_func)
     sort_notes(notes, orderby)
     return notes
 
@@ -588,17 +618,11 @@ def list_by_date(field, creator, date):
 
 @xutils.timeit(name = "NoteDao.CountNote", logfile=True, logargs=True, logret=True)
 def count_by_creator(creator):
-    if xconfig.DB_ENGINE == "sqlite":
-        db    = xtables.get_file_table()
-        where = "is_deleted = 0 AND creator = $creator AND type != 'group'"
-        count = db.count(where, vars = dict(creator = xauth.get_current_name()))
-    else:
-        def count_func(key, value):
-            if value.is_deleted:
-                return False
-            return value.creator == creator and type != 'group'
-        count = dbutil.prefix_count("note_tiny:%s" % creator, count_func)
-    return count
+    def count_func(key, value):
+        if value.is_deleted:
+            return False
+        return value.creator == creator and type != 'group'
+    return dbutil.prefix_count("note_tiny:%s" % creator, count_func)
 
 def count_user_note(creator):
     return count_by_creator(creator)
@@ -620,11 +644,13 @@ def count_by_parent(creator, parent_id):
 
     return dbutil.prefix_count("note_tiny", list_note_func)
 
+@xutils.timeit(name = "NoteDao.CountDict", logfile = True, logargs = True, logret = True)
+def count_dict(user_name):
+    import xtables
+    return xtables.get_dict_table().count()
+
 @xutils.timeit(name = "NoteDao.FindPrev", logfile = True)
 def find_prev_note(note, user_name):
-    # where = "parent_id = $parent_id AND name < $name ORDER BY name DESC LIMIT 1"
-    # table = xtables.get_file_table()
-    # return table.select_first(where = where, vars = dict(name = note.name, parent_id = note.parent_id))
     parent_id = str(note.parent_id)
     note_name = note.name
     def find_prev_func(key, value):
@@ -641,9 +667,6 @@ def find_prev_note(note, user_name):
 
 @xutils.timeit(name = "NoteDao.FindNext", logfile = True)
 def find_next_note(note, user_name):
-    # where = "parent_id = $parent_id AND name > $name ORDER BY name ASC LIMIT 1"
-    # table = xtables.get_file_table()
-    # return table.select_first(where = where, vars = dict(name = note.name, parent_id = note.parent_id))
     parent_id = str(note.parent_id)
     note_name = note.name
     def find_next_func(key, value):
@@ -657,13 +680,6 @@ def find_next_note(note, user_name):
         return result[0]
     else:
         return None
-
-def update_priority(creator, id, value):
-    table = xtables.get_file_table()
-    rows = table.update(priority = value, where = dict(creator = creator, id = id))
-    cache_key = "[%s]note.recent" % creator
-    cacheutil.prefix_del(cache_key)
-    return rows > 0
 
 def add_history(id, version, note):
     if version is None:
@@ -732,7 +748,17 @@ def list_removed(creator, offset, limit):
     return notes
 
 def doc_filter_func(key, value):
-    return value.type not in ("csv", "gallery", "list", "table", "group") and value.is_deleted == 0
+    return value.type in ("md", "text", "html", "post") and value.is_deleted == 0
+
+def table_filter_func(key, value):
+    return value.type in ("csv", "table") and value.is_deleted == 0
+
+def get_filter_func(type, default_filter_func):
+    if type == "document" or type == "doc":
+        return doc_filter_func
+    if type in ("csv", "table"):
+        return table_filter_func
+    return default_filter_func
 
 def list_by_type(creator, type, offset, limit, orderby = "name", skip_archived = False):
     def list_func(key, value):
@@ -740,21 +766,16 @@ def list_by_type(creator, type, offset, limit, orderby = "name", skip_archived =
             return False
         return value.type == type and value.creator == creator and value.is_deleted == 0
 
-    if type == "document" or type == "doc":
-        list_func = doc_filter_func
-
-    notes = dbutil.prefix_list("note_tiny:%s" % creator, list_func, offset, limit, reverse = True)
+    filter_func = get_filter_func(type, list_func)
+    notes = dbutil.prefix_list("note_tiny:%s" % creator, filter_func, offset, limit, reverse = True)
     sort_notes(notes, orderby)
     return notes
 
 def count_by_type(creator, type):
-    def base_count_func(key, value):
+    def default_count_func(key, value):
         return value.type == type and value.creator == creator and value.is_deleted == 0
-    if type == "doc":
-        count_func = doc_filter_func
-    else:
-        count_func = base_count_func
-    return dbutil.prefix_count("note_tiny:%s" % creator, count_func)
+    filter_func = get_filter_func(type, default_count_func)
+    return dbutil.prefix_count("note_tiny:%s" % creator, filter_func)
 
 def list_sticky(creator, offset = 0, limit = 1000):
     def list_func(key, value):
@@ -882,8 +903,10 @@ def refresh_note_stat(user_name):
     stat.doc_count     = count_by_type(user_name, "doc")
     stat.gallery_count = count_by_type(user_name, "gallery")
     stat.list_count    = count_by_type(user_name, "list")
-    stat.table_count   = count_by_type(user_name, "csv")
+    stat.table_count   = count_by_type(user_name, "table")
     stat.sticky_count  = count_sticky(user_name)
+    stat.removed_count = count_removed(user_name)
+    stat.dict_count    = count_dict(user_name)
 
     dbutil.put("user_stat:%s:note" % user_name, stat)
     return stat
@@ -904,11 +927,12 @@ xutils.register_func("note.visit",  visit_note)
 xutils.register_func("note.delete", delete_note)
 xutils.register_func("note.touch",  touch_note)
 xutils.register_func("note.update_tags", update_tags)
-xutils.register_func("note.update_priority", update_priority)
+xutils.register_func("note.create_token", create_token)
 
 # query functions
 xutils.register_func("note.get_root", get_root)
 xutils.register_func("note.get_by_id", get_by_id)
+xutils.register_func("note.get_by_token", get_by_token)
 xutils.register_func("note.get_by_id_creator", get_by_id_creator)
 xutils.register_func("note.get_by_name", get_by_name)
 xutils.register_func("note.get_tags", get_tags)
